requirements.txt
google-api-python-client>=2.131.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.1
